migration (#722)

+++ crates/schema/src/models.rs
@@ -684,7 +684,10 @@ pub struct MarginManagerState {
     pub created_at: chrono::NaiveDateTime,
     #[serde(serialize_with = "serialize_datetime")]
     pub updated_at: chrono::NaiveDateTime,
-    pub current_price: Option<i64>,
-    pub lowest_trigger_above_price: Option<i64>,
-    pub highest_trigger_below_price: Option<i64>,
+    #[serde(serialize_with = "serialize_bigdecimal_option")]
+    pub current_price: Option<BigDecimal>,
+    #[serde(serialize_with = "serialize_bigdecimal_option")]
+    pub lowest_trigger_above_price: Option<BigDecimal>,
+    #[serde(serialize_with = "serialize_bigdecimal_option")]
+    pub highest_trigger_below_price: Option<BigDecimal>,
 }
